lib/tests: Janitor patch
Fix a bunch of compiler warnings.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2744 5f96b256-904b-4d8d-8c98-d829582c6739

+++ lib/test-iobroker.c
@@ -60,13 +60,13 @@ static int connected_handler(int fd, int events, void *arg)
 
 		buf[len] = 0;
 
-		test(len == strlen(msg[i]), "len match for message %d", i);
+		test(len == (int)strlen(msg[i]), "len match for message %d", i);
 		test(!memcmp(buf, msg[i], len), "data match for message %d", i);
 	}
 
 	i++;
 
-	if (i < 0 || i >= ARRAY_SIZE(msg)) {
+	if (i < 0 || i >= (int)ARRAY_SIZE(msg)) {
 		fprintf(stderr, "i = %d in connected_handler(). What's up with that?\n", i);
 		return 0;
 	}
